% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_simresults.R
\name{check_simresults}
\alias{check_simresults}
\title{A function that checks if a call to simulation function worked}
\usage{
check_simresults(simresult)
}
\arguments{
\item{simresult}{a list of results from a single call to the simulation functions}
}
\value{
Ane error message string or NULL
}
\description{
This function takes the return from a call to a simulate_
function and checks that no errors occured.
If something goes wrong, an error message is returned
}
\details{
This function checks simulation results generated from a single call to a simulate_ function.
}
