% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_appsettings.R
\name{construct_appsettings}
\alias{construct_appsettings}
\title{A function to parse appsettings}
\usage{
construct_appsettings(appName, at, modeldir, simdir)
}
\arguments{
\item{appName}{an app name}

\item{modeldir}{model directory}

\item{simdir}{sim directory}
}
\value{
the information is stored in a list called 'appsettings' \cr
different models can have different variables \cr
all models need the following: \cr
variable appid - ID (short name) of the app \cr
variable apptitle - the name of the app. Used to display. \cr
variable docname - name of documentation file for app \cr
variable modelfigname - name of figure file for app \cr
variable simfunction - the name of the simulation function(s) \cr
variable mbmodelname - if there is an mbmodel available, list its name \cr
variable modeltype - the type of the model to be run. if multiple, i.e. containing "_and_" it is set by UI. \cr
 \cr
additional elements that can be provided: \cr
variable otherinputs - contains additional shiny UI elements that are not generated automatically by functions above \cr
for instance all non-numeric inputs need to be provided separately. \cr
this is provided as text \cr
If not needed, it is empty "" \cr
}
\description{
This function parses the apptable for a specific appName preparing for generate_shinyinput()
}
\details{
This function returns specific settings for apps.
}
