% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsairmapps.R
\name{dsairmapps}
\alias{dsairmapps}
\title{A helper function that returns a list of existing apps/simulations}
\usage{
dsairmapps()
}
\description{
This function opens returns a list of existing apps
}
\details{
Run this function with no arguments to list all apps.
}
\examples{
# To see all available apps, run
dsairmapps()
}
\author{
Andreas Handel
}
