% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsairmmenu.R
\name{dsairmmenu}
\alias{dsairmmenu}
\title{The main menu for the DSAIRM package}
\usage{
dsairmmenu()
}
\description{
This function opens a Shiny App menu that will allow the user to run the different simulation apps
}
\details{
Run this function with no arguments to start the main menu
}
\examples{
\dontrun{dsairmmenu()}
}
\author{
Andreas Handel
}
