% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_drugresistance.r
\name{evolutionratefunc}
\alias{evolutionratefunc}
\title{Stochastic simulation of a compartmental acute virus infection model
with treatment and drug resistant strain}
\usage{
evolutionratefunc(y, parms, t)
}
\arguments{
\item{U0}{initial number of taget cells}

\item{Is0}{initial number of wild-type infected cells}

\item{Ir0}{initial number of resistant infected cells}

\item{Vs0}{initial number of wild-type virus}

\item{Vr0}{initial number of resistant virus}

\item{b}{level/rate of infection of cells}

\item{dI}{rate of infected cell death}

\item{e}{efficacy of drug}

\item{m}{fraction of resistant mutants created}

\item{p}{virus production rate}

\item{c}{virus removal rate}

\item{f}{fitness cost of resistant virus}

\item{tmax}{maximum simulation time, units depend on choice of units for your
parameters}
}
\value{
This function returns the simulation result as obtained from a call
  to the adaptivetau integrator
}
\description{
Simulation of a stochastic model with the following compartments:
Uninfected target cells (U), Infected with wild-type/sensitive and untreated (Is),
infected with resistant (Ir), wild-type virus (Vs), resistant virus (Vr)
}
\details{
A compartmental ID model with several states/compartments
is simulated as a stochastic model using the adaptive tau algorithm as implemented by ssa.adaptivetau
in the adpativetau package. See the manual of this package for more details.
The function returns the time series of the simulated disease as output matrix,
with one column per compartment/variable. The first column is time.
}
\section{Warning}{

This function does not perform any error checking. So if you try to do
something nonsensical (e.g. have I0 > PopSize or any negative values or fractions > 1),
the code will likely abort with an error message
}

\examples{
# To run the simulation with default parameters just call this function
result <- simulate_resistance()
# To choose parameter values other than the standard one, specify them e.g. like such
result <- simulate_resistance(tmax = 200, e = 0.5)
# You should then use the simulation result returned from the function, e.g. like this:
plot(result[,1],result[,2],xlab='Time',ylab='Uninfected cells',type='l')
}
\references{
See the manual for the adaptivetau package for details on the algorithm.
            The implemented model is loosely based on: Handel et al 2007 PLoS Comp Bio
           "Neuraminidase Inhibitor Resistance in Influenza: Assessing the Danger of Its
           Generation and Spread"
}
\author{
Andreas Handel
}
