% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_equations.R
\name{generate_equations}
\alias{generate_equations}
\title{Turn a model into a set of differential equations displayed as LaTeX/HTML object}
\usage{
generate_equations(mbmodel)
}
\arguments{
\item{mbmodel}{modelbuilder model structure, either as list object or file name}
}
\value{
The function returns equations as an html object
}
\description{
The model needs to adhere to the structure specified by the modelbuilder package
models built using the modelbuilder package automatically have the right structure
a user can also build a model list structure themselves following the specifications
if the user provides a file name, this file needs to contain an object called 'model'
and contain a valid modelbuilder model structure
}
\details{
This function takes as input a model and produces output that displays ODE equations
}
\author{
Andreas Handel
}
