% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fctcall.R
\name{generate_fctcall}
\alias{generate_fctcall}
\title{A helper function that takes a model and settings and generates the function call needed to run it}
\usage{
generate_fctcall(modelsettings, modelfunction)
}
\arguments{
\item{modelsettings}{values for model settings}

\item{modelfunction}{a name of the model function to be run for the indicated settings}
}
\value{
a string that can be evaluated to run the model
}
\description{
This function generates a text string that contains the function call for a given model
This is a helper function.
}
\details{
This function produces a function text string.
If modelfunction is a modelbuilder object, it will be parsed.
If modelfunction is a character string, this is the function to be run.
}
\author{
Andreas Handel
}
