% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fctcalls.R
\name{generate_fctcalls}
\alias{generate_fctcalls}
\title{A helper function that produces call(s) to a simulator function for specific settings}
\usage{
generate_fctcalls(modelsettings)
}
\arguments{
\item{modelsettings}{a list with model settings. \cr
Generally, this list is produced by calling generate_modelsettings() \cr
Required list elements are: \cr
modelsettings$simfunction - name of simulation function(s) as string.  \cr
modelsettings$modeltype - specify what kind of model should be run.
Currently one of: _ode_, _discrete_, _stochastic_, _usanalysis_, _modelexploration_, _fit_. \cr
For more than one model type, place _and_ between them. \cr
modelsettings$plottype - 'Boxplot' or 'Scatterplot' , required for US app \cr
Optional list elements are: \cr
List elements with names and values for inputs expected by simulation function.
If not provided, defaults of simulator function are used.\cr
modelsettings$plotscale - indicate which axis should be on a log scale (x, y or both).
If not provided or set to '', no log scales are used. \cr
modelsettings$nplots -  indicate number of plots that should be produced (number of top list elements in result).
If not provided, a single plot is assumed.  \cr
modelsettings$nreps - required for stochastic models to indicate numer of repeat simulations.
If not provided, a single run will be done. \cr}
}
\value{
A list containing one or multiple unevaluated function calls with the specified settings, or string with an error message
}
\description{
This function takes a modelsettings structure and uses that information
to create an unevaluated function call that runs the simulator function with the specified settings
}
\details{
This function produces function call(s) for specific settings.
}
