% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_instruction_tabs.R
\name{generate_instruction_tabs}
\alias{generate_instruction_tabs}
\title{A helper function for the UI part of the shiny apps}
\usage{
generate_instruction_tabs()
}
\value{
tablist a list of tabs for display in a shiny UI
}
\description{
This function generates the tabs with content for each Shiny app. This is a helper function.
}
\details{
This function is called by the shiny UIs to populate the documentation and information tabs
}
\author{
Andreas Handel
}
