% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_output.R
\name{generate_output}
\alias{generate_output}
\title{A function that takes shiny GUI input and generates intermediate code used
to run simulations and process results}
\usage{
generate_output(modelsettings, resultslist)
}
\arguments{
\item{modelsettings}{a list with model settings. Required list elements are: \cr
modelsettings$simfunction - name of simulation function(s) as string.  \cr
modelsettings$is_mbmodel - indicate of simulation function has mbmodel structure
modelsettings$modeltype - specify what kind of model should be run.
Currently one of: _ode_, _discrete_, _stochastic_, _usanalysis_, _modelexploration_, _fit_. \cr
For more than one model type, place _and_ between them. \cr
modelsettings$plottype - 'Boxplot' or 'Scatterplot' , required for US app \cr
Optinal list elements are: \cr
List elements with names and values for inputs expected by simulation function.
If not provided, defaults of simulator function are used.\cr
modelsettings$plotscale - indicate which axis should be on a log scale (x, y or both).
If not provided or set to '', no log scales are used. \cr
modelsettings$nplots -  indicate number of plots that should be produced (number of top list elements in result).
If not provided, a single plot is assumed.  \cr
modelsettings$nreps - required for stochastic models to indicate numer of repeat simulations.
If not provided, a single run will be done. \cr}

\item{resultslist}{a list of results corresponding to individual calls to simulation functions generated from the provided modelsettings}
}
\value{
A vectored list named "result" with each main list element containing the simulation results in a dataframe called dat and associated metadata required for generate_plot and generate_text functions. Most often there is only one main list entry (result[[1]]) for a single plot/text.
}
\description{
This function is based on run_model() but instead of running the models
it outputs code equivalent to DSAIRM back-end, server processes initiated from shiny GUI.
}
\details{
This function processes and aggregates simulation results generated from a single set of modelsettings.
}
