% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{generate_results}
\alias{generate_results}
\title{A function that takes a list of raw simulation results
and processes them}
\usage{
generate_results(simlist, modelsettings)
}
\arguments{
\item{simlist}{a nested list of results from call(s) to the simulation functions}

\item{modelsettings}{a list with model settings information, usually from generate_modelsettings()}
}
\value{
A vectored list named "result" with each main list element containing the simulation results in a dataframe called dat and associated metadata required for generate_plot and generate_text functions. Most often there is only one main list entry (result[[1]]) for a single plot/text.
}
\description{
This function processes simulation results generated by
calling one or multiple simulation functions.
}
\details{
This function processes and aggregates simulation results generated from a single set of modelsettings.
}
