% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simulationcode.R
\name{generate_simulationcode}
\alias{generate_simulationcode}
\title{A function that takes shiny GUI input and generates intermediate code used
to run simulations}
\usage{
generate_simulationcode(modelsettings)
}
\arguments{
\item{modelsettings}{a list with model settings. Required list elements are: \cr
modelsettings$simfunction - name of simulation function(s) as string.  \cr
modelsettings$is_mbmodel - indicate of simulation function has mbmodel structure
modelsettings$modeltype - specify what kind of model should be run.
Currently one of: _ode_, _discrete_, _stochastic_, _usanalysis_, _modelexploration_, _fit_. \cr
For more than one model type, place _and_ between them. \cr
modelsettings$plottype - 'Boxplot' or 'Scatterplot' , required for US app \cr
Optinal list elements are: \cr
List elements with names and values for inputs expected by simulation function.
If not provided, defaults of simulator function are used.\cr
modelsettings$plotscale - indicate which axis should be on a log scale (x, y or both).
If not provided or set to '', no log scales are used. \cr
modelsettings$nplots -  indicate number of plots that should be produced (number of top list elements in result).
If not provided, a single plot is assumed.  \cr
modelsettings$nreps - required for stochastic models to indicate numer of repeat simulations.
If not provided, a single run will be done. \cr}
}
\value{
A list with two elements: the first is the modelsettings$modeltype; the second is a nested list of function calls generated according to the modelsettings.
}
\description{
This function takes a modelsettings list is based on run_model() but instead of running the models
it outputs code equivalent to DSAIRM back-end, server processes initiated from shiny GUI.
}
\details{
This function provides a list of function calls necessary in constructing code for download from the shiny GUI.
}
