% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simulationscript.R
\name{generate_simulationscript}
\alias{generate_simulationscript}
\title{A function that prepares simulator code to write file for code download}
\usage{
generate_simulationscript(modelsettings)
}
\arguments{
\item{modelsettings}{A list with model settings.
#' Needs to contain list elements with names and values for all inputs expected
by simulation function. Assumed to be the return of generate_modelsettings()}
}
\value{
Creates a character vector for an R script that runs the simulation specified in the app
and returns the text and plots created by the simulation.
}
\description{
This function is called when the user clicks the "Download Code" button in
the simulator app. It creates an R script containing the code to run the
simulation specified in the app from which the button is clicked and to generate output
similar to what is shown in the app.
}
