% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_warnings.R
\name{generate_warnings}
\alias{generate_warnings}
\title{A helper function that takes result from the simulators and produces warnings}
\usage{
generate_warnings(input, output)
}
\arguments{
\item{input}{the shiny app input structure}

\item{output}{the shiny app output structure}
}
\value{
output warn elements for display in a shiny UI
}
\description{
This function generates warnings to be displayed in the Shiny UI.
This is a helper function. This function processes results returned from the simulation
}
\details{
This function is called by the shiny server to produce output returned to the shiny UI
}
\author{
Andreas Handel
}
