% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{A function that runs a DSAIRM/DSAIDE app}
\usage{
run_model(modelsettings, modelfunction)
}
\arguments{
\item{modelsettings}{vector of model settings. needs to contain inputs expected by simulation function. Also needs to provide further information.}

\item{modelfunction}{The name of a simulation function to be run with the indicated settings.}
}
\value{
A list named "result" with the simulated dataframe and associated metadata.
}
\description{
This function takes a model and model settings and runs it.
It runs the simulation determined by the model settings and returns simulation results.
}
\details{
This function runs a model for specific settings. It is similar to analyze_model in the modelbuilder package.
}
\author{
Andreas Handel
}
