% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_basicmodelfit.R
\name{simulate_basicmodelfit}
\alias{simulate_basicmodelfit}
\title{Fitting a simple viral infection models to influenza data}
\usage{
simulate_basicmodelfit(U0 = 1e+05, I0 = 0, V0 = 1, X0 = 1, n = 0,
  dU = 0, dI = 1, p = 10, g = 1, b = 1e-05, bsim = 1e-05,
  blow = 1e-06, bhigh = 0.001, dV = 2, dVsim = 2, dVlow = 0.001,
  dVhigh = 1000, usesimdata = 1, noise = 0, iter = 100)
}
\arguments{
\item{U0}{initial number of uninfected target cells}

\item{I0}{initial number of infected target cells}

\item{V0}{initial number of infectious virions}

\item{X0}{initial level of immune response}

\item{dI}{rate at which infected cells die}

\item{p}{rate at which infected cells produce virus}

\item{b}{rate at which virus infects cells}

\item{blow}{lower bound for infection rate}

\item{bhigh}{upper bound for infection rate}

\item{dV}{rate at which infectious virus is cleared}

\item{iter}{max number of steps to be taken by optimizer}
}
\value{
The function returns a list containing the best fit timeseries, the best fit parameters, and AICc for the model
}
\description{
This function runs a simulation of a compartment model
using a set of ordinary differential equations.
The model describes a simple viral infection system
}
\details{
a simple compartmental ODE model mimicking acute viral infection
is fitted to data
Data can either be real or created by running the model with known parameters and using the simulated data to
determine if the model parameters can be identified
}
\section{Warning}{
 This function does not perform any error checking. So if
  you try to do something nonsensical (e.g. specify negative parameter or starting values,
  the code will likely abort with an error message
}

\examples{
# To run the code with default parameters just call this function
result <- simulate_basicmodelfit()
# To choose parameter values other than the standard one, specify them e.g. like such
result <- simulate_basicmodelfit(U0 = 1e6, dI = 2)
}
\seealso{
See the shiny app documentation corresponding to this
function for more details on this model.
}
\author{
Andreas Handel
}
