% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_basicvirus_stochastic.R
\name{simulate_basicvirus_stochastic}
\alias{simulate_basicvirus_stochastic}
\title{Stochastic simulation of an SEIR-type model}
\usage{
simulate_basicvirus_stochastic(U0 = 10000, I0 = 0, V0 = 10, tmax = 100,
  n = 0, dU = 0, b = 0.001, dI = 1, p = 2, dV = 2)
}
\arguments{
\item{U0}{initial number of uninfected cells}

\item{I0}{initial number of infected cells}

\item{V0}{initial number of virions}

\item{tmax}{maximum simulation time, units depend on choice of units for
your parameters}

\item{n}{rate of uninfected cell production}

\item{dU}{rate of uninfected cell death (the inverse is the average lifespan)}

\item{b}{rate of cell infection}

\item{dI}{rate of infected cell death (the inverse is the average lifespan)}

\item{p}{rate of new virus production}

\item{dV}{rate of virus removal (the inverse is the average lifespan)}
}
\value{
The function returns the time series of the simulated model as
  matrix, with one column per compartment/variable. The first column is time.
}
\description{
Simulation of a stochastic simple virus infection model with the following
  compartments: Uninfected cells (U), Infected cells (I), free virus (V)
}
\details{
A compartmental, stochastic model with several states/compartments is
  simulated.  Units of time depend on the time units chosen for model parameters.
  The simulation runs as a stochastic model using the adaptive-tau algorithm as implemented
  by ssa.adaptivetau in the adpativetau package. See the manual of this
  package for more details. The function returns the time series of the
  simulated disease as output matrix, with one column per
  compartment/variable. The first column is time.
}
\section{Warning}{
 This function does not perform any error checking. So if
  you try to do something nonsensical (e.g. specify negative parameter values
  or fractions > 1), the code will likely abort with an error message
}

\examples{
# To run the simulation with default parameters
result <- simulate_basicvirus_stochastic()
# To choose parameter values other than the standard one, specify them e.g. like this
result <- simulate_basicvirus_stochastic(V0 = 100,  tmax = 200, dI = 2)

# You can display or further process the result, e.g. like this
plot(result[,'time'],result[,'V'],xlab='Time',ylab='Virus',type='l')
print(paste('Peak number of virus:',max(result[,'V'])))
}
\seealso{
See the shiny app documentation corresponding to this simulator
function for more details on this model. See the manual for the adaptivetau
package for details on the stochastic algorithm.
}
\author{
Andreas Handel
}
