% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_usanalysis.R
\name{simulate_usanalysis}
\alias{simulate_usanalysis}
\title{Simulation to do an uncertainty and sensitivity analysis using the simple bacteria model}
\usage{
simulate_usanalysis(B0min = 10, B0max = 10, I0min = 1, I0max = 1,
  Bmaxmin = 1e+06, Bmaxmax = 1e+06, dBmin = 0.1, dBmax = 0.1,
  kmin = 1e-07, kmax = 1e-07, rmin = 0.001, rmax = 0.001, dImin = 1,
  dImax = 2, gmean = 0.5, gvar = 0.1, tmax = 30, samples = 10)
}
\arguments{
\item{B0min}{lower bound for initial bacteria numbers}

\item{B0max}{upper bound for initial bacteria numbers}

\item{I0min}{lower bound for initial immune response}

\item{I0max}{upper bound for initial immune response}

\item{Bmaxmin}{lower bound for maximum bacteria load}

\item{Bmaxmax}{upper bound for maximum bacteria load}

\item{dBmin}{lower bound for bacteria death rate}

\item{dBmax}{upper bound for bacteria death rate}

\item{kmin}{lower bound for immune response kill rate}

\item{kmax}{upper bound for immune response kill rate}

\item{rmin}{lower bound for immune response growth rate}

\item{rmax}{upper bound for immune response growth rate}

\item{dImin}{lower bound for immune response death rate}

\item{dImax}{upper bound for immune response death rate}

\item{gmean}{mean for bacteria growth rate}

\item{gvar}{variance for bacteria growth rate}

\item{tmax}{maximum simulation time, units depend on choice of units for your
parameters}

\item{samples}{number of LHS samples to run}
}
\value{
The function returns the output as a data frame,
with sample values for each parameter as columns, followed by columns for the results.
}
\description{
This function performs latin hypercube sampling of parameters
for the simple bacteria model underlying the app by the same name
The user provides ranges for the initial conditions and parameter values and the number of samples.
The function does LHS of the ODE using an ODE solver from the deSolve package.
The function returns a data frame containing values for each sample and results
}
\details{
A simple 2 compartment ODE model (the simple bacteria model introduced in the app of that name)
is simulated for different parameter values.
Parameters are sampled via latin hypercube sampling.
Distribution for all parameters is assumed to be uniform between the min and max values.
The only exception is the bacteria growth parameter,
which is assumed to be gamma distributed with the specified mean and variance
the simulation returns for each parameter sample the peak and final value for B and I
also returned are all parameter values as individual columns
}
\section{Warning}{
 This function does not perform any error checking. So if
  you try to do something nonsensical (e.g. specify negative parameter values
  or fractions > 1), the code will likely abort with an error message
}

\examples{
# To run the simulation with default parameters just call this function
result <- simulate_usanalysis()
# To choose parameter values other than the standard one, specify them e.g. like such
result <- simulate_usanalysis(dImin = 0.1, dImax = 10, tmax = 100)
# You should then use the simulation result returned from the function, e.g. like this:
plot(result[,"dI"],result[,"Bpeak"],xlab='values for d',ylab='Peak Bacteria Number',type='l')
}
\seealso{
See the shiny app documentation corresponding to this simulator
function for more details on this model.
}
\author{
Andreas Handel
}
