% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAIRM-package.R
\docType{package}
\name{DSAIRM-package}
\alias{DSAIRM}
\alias{DSAIRM-package}
\title{DSAIRM: A package to learn about Dynamical Systems Approaches to
Immune Response Modeling}
\description{
This package provides a number of Shiny apps that simulate various
within-host infection and immune response dynamics models.
By manipulating the models and working
through the instructions provided within the Shiny UI, you can learn about
some important concepts in immune response modeling. You will also
learn how models can be used to study such concepts.
}
\details{
Start the main menu of the package by calling dsairmmenu().

To learn more about how to use the package, see the documentation
on the package website:
https://ahgroup.github.io/DSAIRM/
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ahgroup.github.io/DSAIRM/}
  \item \url{https://github.com/ahgroup/DSAIRM/}
  \item Report bugs at \url{https://github.com/ahgroup/DSAIRM/issues}
}

}
\author{
\strong{Maintainer}: Andreas Handel \email{andreas.handel@gmail.com} (\href{https://orcid.org/0000-0002-4622-1146}{ORCID})

Other contributors:
\itemize{
  \item Cody Dailey [contributor]
  \item Yang Ge [contributor]
  \item Spencer Hall [contributor]
  \item Brian McKay [contributor]
  \item Sina Solaimanpour [contributor]
  \item Alexis Vittengl [contributor]
  \item Henok Woldu [contributor]
}

}
\keyword{internal}
