% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fctcall.R
\name{generate_fctcall}
\alias{generate_fctcall}
\title{A helper function that produces a call to a simulator function for specific settings}
\usage{
generate_fctcall(modelsettings)
}
\arguments{
\item{modelsettings}{a list with model settings. Required list elements are: \cr
List elements with names and values for all inputs expected by simulation function. \cr
modelsettings$simfunction - name of simulation function in variable  \cr}
}
\value{
A string containing an unevaluated function call with the specified settings
}
\description{
This function takes a modelsettings structure and uses that information
to create an unevaluated function call that runs the simulator function with the specified settings
}
\details{
This function produces a function call for specific settings.
}
