% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{generate_plots}
\alias{generate_plots}
\title{A helper function that takes result from the simulators and produces plots}
\usage{
generate_plots(input, output, allres)
}
\arguments{
\item{input}{the shiny app input structure}

\item{output}{the shiny app output structure}

\item{allres}{a list containing all simulation results.
the length of the list indicates the number of separate plots to make
each list entry corresponds to one plot
each list entry needs to contain the following information
1. a data frame called dat with one column called xvals, one column yvals,
one column called varnames that contains names for different variables
and one column called IDvar for further grouping (i.e. multiple lines for stochastic simulations)
supplying IDvar is optional, if it is missing, this grouping is ignored
2. meta-data for the plot, provided in the following variables:
plottype - one of "Lineplot","Scatterplot","Boxplot"
xlab, ylab - strings to label axes
xscale, yscale - scaling of axes, e.g. "identity" or "log10"
legend - string for legend title
linesize - width of line, numeric, i.e. 1.5, 2, etc. set to 1.5 if not supplied}
}
\value{
output a list with plots for display in a shiny UI
}
\description{
This function generates plots to be displayed in the Shiny UI.
This is a helper function. This function processes results returned from the simulation, supplied as a list
}
\details{
This function is called by the shiny server to produce plots returned to the shiny UI
}
\author{
Andreas Handel
}
