% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{generate_plots}
\alias{generate_plots}
\title{A helper function that takes result from the simulators and produces plots}
\usage{
generate_plots(input, output, allres)
}
\arguments{
\item{input}{the shiny app input structure}

\item{output}{the shiny app output structure}

\item{allres}{a list containing all simulation results.
the length of the list indicates number of separate plots to make, each list entry is one plot and one set of text output
for each list entry (e.g. allres[[i]]$type), 'type' specifies the kind of plot
allres[[i]]$dat contains a dataframe in tidy/ggplot format for plotting. one column is called xvals, one column yvals, further columns are stratifiers/aesthetics, e.g. names of variables or number of run for a given variable}
}
\value{
output a list with plots for display in a shiny UI
}
\description{
This function generates plots to be displayed in the Shiny UI.
This is a helper function. This function processes results returned from the simulation, supplied as a list
}
\details{
This function is called by the shiny server to produce plots returned to the shiny UI
}
\author{
Andreas Handel
}
