% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shinyinput.R
\name{generate_shinyinput}
\alias{generate_shinyinput}
\title{A helper function that takes a model and generates the shiny UI elements for the analyze tab}
\usage{
generate_shinyinput(mbmodel, otherinputs = NULL, packagename)
}
\arguments{
\item{mbmodel}{a name of a function for which to build inputs}

\item{otherinputs}{a list of other shiny inputs to include}

\item{packagename}{name of package using this function}
}
\value{
A renderUI object that can be added to the shiny output object for display in a Shiny UI
}
\description{
This function generates numeric shiny UI inputs for a supplied model.
This is a helper function called by the shiny app.
}
\details{
This function is called by the Shiny app to produce the Shiny input UI elements.
mbmodel is assumed to be the name of a function which will be parsed to create UI elements.
Non-numeric arguments of functions are removed.
}
\author{
Andreas Handel
}
