% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shinyinput.R
\name{generate_shinyinput}
\alias{generate_shinyinput}
\title{A helper function that takes a model and generates the shiny UI elements for the analyze tab}
\usage{
generate_shinyinput(mbmodel, otherinputs = NULL, output)
}
\arguments{
\item{mbmodel}{a modelbuilder model structure}

\item{otherinputs}{additional elements to be shown in the UI}

\item{output}{shiny output structure}
}
\value{
No direct return. output structure is modified to contain text for display in a Shiny UI
}
\description{
This function generates input buttons and sliders for a supplied model.
This is a helper function called by the shiny app.
}
\details{
This function is called by the Shiny app to produce the Shiny input UI elements.
}
\author{
Andreas Handel
}
