% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_text.R
\name{generate_text}
\alias{generate_text}
\title{A helper function that takes result from the simulators and produces text output}
\usage{
generate_text(input, output, allres)
}
\arguments{
\item{input}{the shiny app input structure, mainly as a pass-through}

\item{output}{the shiny app output structure, mainly as a pass-through}

\item{allres}{a list containing all simulation results and other information
if the list entry 'maketext'  exists for a given plot (list element in allres)
the data needs to follow the structure described in the generate_plots function
if 'maketext' is false, this function only uses - if present' the entries 'showtext'
for each plot and finaltext of the 1st list elemnt for an overall message/text
see the generate_plots function for more details on the structure of the list}
}
\value{
HTML formatted text for display in a shiny UI
}
\description{
This function generates text to be displayed in the Shiny UI.
This is a helper function. This function processes results returned from the simulation, supplied as a list
}
\details{
This function is called by the shiny server to produce output returned to the shiny UI
}
\author{
Andreas Handel
}
