% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_text.R
\name{generate_text}
\alias{generate_text}
\title{A helper function that takes result from the simulators and produces text output}
\usage{
generate_text(res)
}
\arguments{
\item{res}{a list structure containing all simulation results that are to be processed
this function is meant to be used together with generate_plots and requires similar information
the length of the list indicates the number of separate plots to make, and for each plot this function produces text
each list entry needs to contain the following information/elements:
1. a data frame called "dat" with one column called xvals, one column yvals,
one column called varnames that contains names for different variables.
varnames needs to be a factor variable or will be converted to one.
Optional, one column called IDvar for further grouping (i.e. multiple lines for stochastic simulations).
If plottype is 'mixedplot' an additional column called 'style' indicating line or point plot
for each variable is needed.
For stochastic simulations that require averaging, an variable called nreps is needed,
which should indicate the number of simulation.
2. meta-data for the plot, provided in the following variables:
optinal: plottype - one of "Lineplot" (chosen if nothing is provided),"Scatterplot","Boxplot", "Mixedplot"
the plottype determines what kind of text is generated.
For lineplots and mixedplot, min/max/final values of each line are shown
for scatterplots, a correlation coefficient is computed
boxplots show min/max/median/average
if the list entry 'maketext' is set to TRUE (or not provided) exists for a given plot the just described outputs will be generated
if 'maketext' is FALSE, no text is generated
if present' the entries 'showtext'
for each plot and finaltext of the 1st list element for an overall message/text
are also shown}
}
\value{
HTML formatted text for display in a shiny UI
}
\description{
This function generates text to be displayed in the Shiny UI.
This is a helper function. This function processes results returned from the simulation, supplied as a list
}
\details{
This function is called by the shiny server to produce output returned to the shiny UI
}
\author{
Andreas Handel
}
