% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_basicbacteria_discrete.R
\name{simulate_basicbacteria_discrete}
\alias{simulate_basicbacteria_discrete}
\title{Basic Bacteria model}
\usage{
simulate_basicbacteria_discrete(B = 10, I = 1, g = 1, Bmax = 1e+06,
  dB = 0.1, k = 1e-07, r = 0.001, dI = 1, tstart = 0,
  tfinal = 30, dt = 0.05)
}
\arguments{
\item{B}{: starting value for Bacteria}

\item{I}{: starting value for Immune Response}

\item{g}{: maximum rate of bacteria growth}

\item{Bmax}{: bacteria carrying capacity}

\item{dB}{: bacteria death rate}

\item{k}{: bacteria kill rate}

\item{r}{: immune response growth rate}

\item{dI}{: immune response decay rate}

\item{tstart}{: Start time of simulation}

\item{tfinal}{: Final time of simulation}

\item{dt}{: Time step}
}
\value{
The function returns the output as a list.
The time-series from the simulation is returned as a dataframe saved as list element \code{ts}.
The \code{ts} dataframe has one column per compartment/variable. The first column is time.
}
\description{
A basic bacteria infection model with 2 compartments
}
\details{
The model includes bacteria and an immune response. The processes are bacteria growth, death and killing, and immune response activation and decay. This is a predator-prey type model.
This code is based on a dynamical systems model created by the modelbuilder package.
The model is implemented here as a set of discrete-time, deterministic equations,
coded as a simple for-loop.
}
\section{Warning}{
 This function does not perform any error checking. So if you try to do something nonsensical (e.g. have negative values for parameters), the code will likely abort with an error message.
}

\section{Model Author}{
 Andreas Handel
}

\section{Model creation date}{
 2018-12-21
}

\examples{
# To run the simulation with default parameters:
result <- simulate_basicbacteria_discrete()
}
